fix git branch switching in test suite
"LIBVIRT_REF" can be set to test against a different libvirt
ref (tag/branch) then 'master'. If you do this with a tag,
and then run the test again without "LIBVIRT_REF" set,
'git pull' will fail as you're not on a branch.

To address this, merely 'fetch' origin instead of pulling,
and then directly access the origin branch, instead of
the local tracking branch.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>